Clean up unneeded code, various other cleanup
