--- core/main_ng.py
@@ -1,55 +1,0 @@
-from termcolor import colored
-
-from core.scenario import Scenario
-from core.helpers import print_ascii_art, get_scenarios_config
-
-
-def select_attack_scenario(cloud_provider):
-    scenarios_config = get_scenarios_config()
-    keys = list(scenarios_config.keys())
-    keys.sort()
-    print(colored('Select Attack Scenario of %s:', color='yellow') % cloud_provider)
-    choices = []
-    for key in keys:
-        index = int(key[-1:])
-        choices.append(index)
-        print(colored('{}. {}: {}'.format(
-            index, scenarios_config[key]['title'], scenarios_config[key]['description']),
-            color='green'))
-    while True:
-        try:
-            choice = int(input(colored('Enter your choice: ', color='yellow')))
-            if choice not in choices:
-                raise ValueError(colored('Invalid choice.', color='red'))
-            return choice
-        except ValueError as e:
-            print(e)
-
-
-def main(cloud_provider, action, simulation, scenario):
-    tool_name = 'C O B R A'
-    print_ascii_art(tool_name)
-    scenario_choice = select_attack_scenario(cloud_provider)
-    scenario = Scenario(scenario_choice)
-    if action == 'launch':
-        if simulation:
-            # TODO: what to do with cloud provider?
-            scenario.setup()
-            scenario.attack()
-            # scenario.destroy()
-            # TODO: ^^^ do we really want to destroy the infra immediately?
-            #   there's a separate command for destroy
-            scenario.generate_report()  # TODO: not implemented
-    elif action == 'status':
-        # TODO
-        # subprocess.call('cd ./scenarios/scenario_2/infra/ && pulumi stack ls', shell=True)
-        pass
-    elif action == 'destroy':
-        scenario.destroy()
-        pass
-    else:
-        print('No options provided. --help to know more')
-
-
-if __name__ == '__main__':
-    main()
