+++ scenarios/README.md
@@ -1,0 +1,1 @@
+These scenarios should be migrated to the `scenarios_ng` directory. When that is complete, this directory should be removed and `scenarios_ng` renamed to `scenarios`.
