+++ scenarios_ng/scenario_0/infra/extra.py
@@ -2,5 +2,5 @@ def deploy_additional_resources():
     pass
 
 
-def destroy_additional_resources():
+def destroy_additional_resources(data):
     pass
